refactor(cli): migrate hazards validators onto discovery
Move the nine validator-only hazards backends (aqueduct, emdat, flodis, flopros,
gdacs, hanze, jrc_flood, nsi, osm) to earthlens.<backend>.cli, register them in
the hazards earthlens.cli table, and drop their entries + functions from core's
_VALIDATORS. Relocate osm/hanze/flodis unit tests to the hazards distribution and
repoint aqueduct's existing provider test. Shrink the guard allow-list by nine.
