+++ libs/core/tests/test_distribution_boundaries.py
@@ -1270,7 +1270,6 @@ class TestCliIsBackendAgnostic:
     #: empty as #863 lands provider by provider).
     PENDING = frozenset(
         {
-            "aqueduct",
             "argo",
             "asf",
             "bathymetry",
@@ -1281,34 +1280,26 @@ class TestCliIsBackendAgnostic:
             "drought",
             "earthdata",
             "ecmwf",
-            "emdat",
             "erddap",
             "eumetsat",
             "fabdem",
             "fdsn",
             "firms",
-            "flodis",
-            "flopros",
             "gbif",
-            "gdacs",
             "gee",
             "ghsl",
             "glaciers",
             "goes",
-            "hanze",
             "hdx",
             "iucn",
             "jaxa",
-            "jrc_flood",
             "mswep",
             "nrel",
-            "nsi",
             "nwm",
             "nwp",
             "obis",
             "openaq",
             "openeo",
-            "osm",
             "pvgis",
             "radar",
             "radklim",
